+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/AccountsController.java
@@ -1,11 +1,11 @@
-package com.CSCI152.team4.server.AccountsReformat;
+package com.CSCI152.team4.server.Accounts;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("api/v1/accounts")
-public class AccountsController2 {
+public class AccountsController {
 
     /**
      * This controller will map to 3 services all related to accounts.
