+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Classes/EmployeeAccount.java
@@ -1,90 +1,104 @@
 package com.CSCI152.team4.server.Accounts.Classes;
 
 import com.CSCI152.team4.server.Accounts.Settings.Permissions;
+
 import javax.persistence.*;
 import java.sql.Timestamp;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-//@Entity
-//@Table(name="Employee_Account")
-public class EmployeeAccount extends WorkerAccount {
+import static java.time.LocalDateTime.now;
 
+@Entity
+@Table(name="Employee_Account")
+public class EmployeeAccount extends WorkerAccount{
 
-    @Id
-    private String accountId;
+    @EmbeddedId
+    AccountId accountId;
 
     @Column(nullable = false)
     @ElementCollection
-    private List<String> permissions_list;
+    private List<String> permissionsList;
 
     public EmployeeAccount() {
     }
 
-    //No permissions passed
-    public EmployeeAccount(String email, String password, String firstName, String lastName,
-                           String jobTitle, Integer businessId) {
-        super(businessId,
-                email,
-                password,
-                firstName,
-                lastName,
-                Timestamp.valueOf(LocalDateTime.now()),
-                jobTitle);
-        this.accountId = UUID.randomUUID().toString();
-        this.permissions_list = new ArrayList<>();
-    }
-
-    //Permissions Passed
-    public EmployeeAccount(String email, String password,
+    public EmployeeAccount(Integer businessId,
+                           String email, String password,
                            String firstName, String lastName,
-                           String jobTitle, List<String> permissions, int businessId) {
-        super(businessId, email, password, firstName, lastName, Timestamp.valueOf(LocalDateTime.now()), jobTitle);
-        this.accountId = UUID.randomUUID().toString();
-        this.setPermissions_list(permissions);
+                           String jobTitle, List<String> permissionsList) {
+        super(businessId, email, password, firstName, lastName, Timestamp.valueOf(now()), jobTitle);
+        this.accountId = new AccountId(UUID.randomUUID().toString(), email, businessId);
+        setPermissionsList(permissionsList);
+    }
+
+    public EmployeeAccount(Integer businessId,
+                           String email, String password,
+                           String firstName, String lastName,
+                           String jobTitle) {
+        super(businessId, email, password, firstName, lastName, Timestamp.valueOf(now()), jobTitle);
+        this.accountId = new AccountId(UUID.randomUUID().toString(), email, businessId);
+        permissionsList = new ArrayList<>();
     }
 
 
-    public void setAccountId(String accountId) {
-        this.accountId = accountId;
+    /**
+     * Set the permissions of the Employee account. Guarantees to ONLY
+     * set valid permissions to prevent security issues.
+     * */
+    public void setPermissionsList(List<String> grantedPermissions){
+        this.permissionsList = new ArrayList<>();
+        if(grantedPermissions == null) return; //cleared list == no permissions
+
+        for(Permissions p : Permissions.values()){
+            if(grantedPermissions.contains(p.toString())){
+                this.permissionsList.add(p.toString());
+            }
+        }
+    }
+
+    public List<String> getPermissionsList(){
+        return this.permissionsList;
+    }
+
+    public void setAccountId(String accountId){
+        this.accountId.setAccountId(accountId);
+    }
+
+    public AccountId getAccountId() { return this.accountId; }
+
+    @Override
+    public String getAccountIdString(){
+        return this.accountId.getAccountId();
+    }
+
+    @Override
+    public void setBusinessId(Integer businessId){
+        if(getBusinessId() == null){
+            super.setBusinessId(businessId);
+            this.accountId.setBusinessId(businessId);
+        }
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        return super.equals(o)
+                && this.getAccountIdString().equals(o);
     }
 
     @Override
     public String toString() {
-        return "EmployeeAccount{" +
-                "accountId='" + accountId + '\'' +
-                ", permissions_list=" + permissions_list +
+        return "AdminAccount{" +
+                "accountId=" + accountId.toString() +
                 ", email='" + this.getEmail() + '\'' +
                 ", businessId='" + this.getBusinessId() + '\'' +
                 ", password='" + this.getPassword() + '\'' +
                 ", firstName='" + this.getFirstName() + '\'' +
                 ", lastName='" + this.getLastName() + '\'' +
                 ", timestamp=" + this.getTimestamp() +
-                ", jobTitle='" + this.getJobTitle()+ '\'' +
+                ", jobTitle='" + this.getJobTitle() + '\'' +
+                ", permissionsList='" + this.permissionsList.toString() + '\'' +
                 '}';
-    }
-
-    public List<String> getPermissions_list() {
-        return permissions_list;
-    }
-
-    public void setPermissions_list(List<String> permissions_list) {
-
-        //clear array list
-        this.permissions_list = new ArrayList<>();
-        if(permissions_list == null) return; //cleared list == no permissions
-        //refill
-        for(Permissions p : Permissions.values()){
-            if(permissions_list.contains(p.toString())){
-                this.permissions_list.add(p.toString());
-            }
-        }
-    }
-
-    @Override
-    public String getAccountId() {
-        return this.accountId;
     }
 }
