--- java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Endpoint/SessionService.java
@@ -1,106 +1,0 @@
-package com.CSCI152.team4.server.Accounts.Endpoint;
-
-import com.CSCI152.team4.server.Accounts.Classes.AdminAccount;
-import com.CSCI152.team4.server.Accounts.Classes.BusinessAccount;
-import com.CSCI152.team4.server.Accounts.Classes.EmployeeAccount;
-import com.CSCI152.team4.server.Accounts.Classes.WorkerAccount;
-import com.CSCI152.team4.server.Accounts.Repos.AdminAccountRepo2;
-import com.CSCI152.team4.server.Accounts.Repos.BusinessAccountRepo2;
-import com.CSCI152.team4.server.Accounts.Repos.EmployeeAccountRepo2;
-import com.CSCI152.team4.server.Util.AccountAuthenticator;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.crypto.bcrypt.BCrypt;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.Optional;
-
-import static org.springframework.security.crypto.bcrypt.BCrypt.checkpw;
-
-//@Service
-public class SessionService {
-
-    private final AdminAccountRepo2 adminAccountRepo2;
-    private final EmployeeAccountRepo2 employeeAccountRepo2;
-    private final BusinessAccountRepo2 businessAccountRepo2;
-    private final AccountAuthenticator accountAuthenticator;
-
-    @Autowired
-    public SessionService(AdminAccountRepo2 adminRepo, EmployeeAccountRepo2 employeeAccountRepo2,
-                          BusinessAccountRepo2 businessAccountRepo2, AccountAuthenticator accountAuthenticator) {
-        this.adminAccountRepo2 = adminRepo;
-        this.employeeAccountRepo2 = employeeAccountRepo2;
-        this.businessAccountRepo2 = businessAccountRepo2;
-        this.accountAuthenticator = accountAuthenticator;
-    }
-
-
-    public WorkerAccount userLogin(WorkerAccount account){
-        //If valid email and password
-        String plainTextPW = account.getPassword();
-        String accountId = setAccountInfoAndReturnId(account);
-        if(verifyLoginCredentials(plainTextPW, account.getPassword()) &&
-                verifyBusinessConnection(account.getBusinessId(), accountId)){
-
-            //Then set token, clear password
-            account.setToken(accountAuthenticator.getToken(accountId));
-            account.setPassword("");
-            //return account info containing token
-            return account;
-        }
-        else{
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    private boolean verifyLoginCredentials(String plainText, String hashed){
-        return isValidPassword(plainText, hashed);
-    }
-
-    private String setAccountInfoAndReturnId(WorkerAccount account){
-
-        Optional<AdminAccount> adminAccount =
-                adminAccountRepo2.findFirstByEmailAndBusinessId(account.getEmail(), account.getBusinessId());
-        Optional<EmployeeAccount> employeeAccount =
-                employeeAccountRepo2.findFirstByEmailAndBusinessId(account.getEmail(), account.getBusinessId());
-
-        WorkerAccount persisted;
-
-        if(adminAccount.isPresent()){
-            persisted = adminAccount.get();
-        }
-        else if(employeeAccount.isPresent()){
-            persisted = employeeAccount.get();
-        }
-        else{
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        }
-
-        account.setPassword(persisted.getPassword());
-
-        return persisted.getAccountId();
-    }
-
-    private boolean isValidPassword(String plainText, String hashedPassword){
-        return BCrypt.checkpw(plainText, hashedPassword);
-    }
-
-    private boolean verifyBusinessConnection(Integer businessId, String accountId){
-
-        Optional<BusinessAccount> optional = businessAccountRepo2.findById(businessId);
-
-        if(optional.isPresent()){
-            BusinessAccount businessAccount = optional.get();
-            return businessAccount.getAdmins().contains(accountId)
-                    || businessAccount.getEmployees().contains(accountId);
-        }
-
-        return false;
-
-
-    }
-    public void userLogout(WorkerAccount account){
-        accountAuthenticator.invalidateToken(account.getToken());
-    }
-
-}
