--- java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Repos/AdminAccountRepo2.java
@@ -1,17 +1,0 @@
-package com.CSCI152.team4.server.Accounts.Repos;
-
-import com.CSCI152.team4.server.Accounts.Classes.AdminAccount;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-//@Repository
-public interface AdminAccountRepo2 extends CrudRepository<AdminAccount, String> {
-
-
-    Optional<AdminAccount> findFirstByEmailAndBusinessId(@Param("email") String email, @Param("businessId") Integer businessId);
-
-}
