+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Repos/BusinessAccountRepo.java
@@ -1,12 +1,10 @@
 package com.CSCI152.team4.server.Accounts.Repos;
 
-import com.CSCI152.team4.server.Accounts.Classes.AdminAccount;
 import com.CSCI152.team4.server.Accounts.Classes.BusinessAccount;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-//@Repository
-public interface BusinessAccountRepo2 extends CrudRepository<BusinessAccount, Integer> {
+@Repository
+public interface BusinessAccountRepo extends CrudRepository<BusinessAccount, Integer> {
 
 }
