+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Repos/RepoManager.java
@@ -1,7 +1,7 @@
-package com.CSCI152.team4.server.AccountsReformat.Repos;
+package com.CSCI152.team4.server.Accounts.Repos;
 
 
-import com.CSCI152.team4.server.AccountsReformat.AccountClasses.*;
+import com.CSCI152.team4.server.Accounts.Classes.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
