+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Requests/BusinessRequest.java
@@ -1,7 +1,7 @@
-package com.CSCI152.team4.server.AccountsReformat.Requests;
+package com.CSCI152.team4.server.Accounts.Requests;
 
-import com.CSCI152.team4.server.AccountsReformat.AccountClasses.AdminAccount;
-import com.CSCI152.team4.server.AccountsReformat.AccountClasses.BusinessAccount;
+import com.CSCI152.team4.server.Accounts.Classes.AdminAccount;
+import com.CSCI152.team4.server.Accounts.Classes.BusinessAccount;
 
 public class BusinessRequest extends AccountCreationRequest {
