+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Requests/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.CSCI152.team4.server.AccountsReformat.Requests;
+package com.CSCI152.team4.server.Accounts.Requests;
 
 public class LoginRequest {
