+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Services/AccountManagementService.java
@@ -1,4 +1,4 @@
-package com.CSCI152.team4.server.AccountsReformat.Services;
+package com.CSCI152.team4.server.Accounts.Services;
 
 import org.springframework.stereotype.Service;
