+++ java services/server/server/src/main/resources/application.properties
@@ -3,7 +3,7 @@
 spring.jpa.properties.hibernate.jdbc.time_zone=UTC
 spring.datasource.url=jdbc:mysql://localhost:3306/frmw
 spring.datasource.username=root
-spring.datasource.password=
+spring.datasource.password=H2vhf794!5377
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=false
