+++ java services/server/server/src/test/java/com/CSCI152/team4/server/Accounts/Services/RegistrationServiceTest.java
@@ -1,10 +1,8 @@
-package com.CSCI152.team4.server.AccountsReformat.Services;
+package com.CSCI152.team4.server.Accounts.Services;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class RegistrationServiceTest {
