--- java services/server/server/src/test/java/com/CSCI152/team4/server/AccountsReformat/Repos/AdminAccountRepoTest.java
@@ -1,22 +1,0 @@
-package com.CSCI152.team4.server.AccountsReformat.Repos;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class AdminAccountRepoTest {
-
-    @Test
-    void itShouldFindTopByEmailAndBusinessId() {
-        // Given
-        // When
-        // Then
-    }
-
-    @Test
-    void itShouldExistsByEmailAndBusinessId() {
-        // Given
-        // When
-        // Then
-    }
-}
