--- java services/server/server/src/test/java/com/CSCI152/team4/server/AccountsReformat/Repos/BusinessAccountRepoTest.java
@@ -1,7 +1,0 @@
-package com.CSCI152.team4.server.AccountsReformat.Repos;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class BusinessAccountRepoTest {
-
-}
